grouped GEMM MoE (#276)

+++ cute_kernels/kernels/grouped_gemm/__init__.py
@@ -26,7 +26,7 @@ def grouped_gemm_cute(
     assert beta == 0
     assert C is None
 
-    output = torch.empty(*output_shape, device=A.device, dtype=torch.bfloat16)
+    output = torch.empty(*output_shape, device=A.device, dtype=A.dtype)
 
     grouped_gemm_cuda(
         A=A,
